requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
datasets>=2.14.0
huggingface-hub>=0.17.0

# LLM Integration
openai>=1.0.0

# Frontend
streamlit>=1.28.0

# Optional: Caching & Storage
redis>=5.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Development
black>=23.0.0
flake8>=6.0.0
